warning fix and square image enforcement removed; plus some code cleanup
