+++ package.json
@@ -38,7 +38,6 @@
     "react-hook-form": "^7.51.5",
     "react-select": "^5.8.0",
     "react-toastify": "^10.0.5",
-    "sharp": "^0.33.4",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
     "unique-username-generator": "^1.3.0",
